Rewrite BufferWorldAllocator
Should be complete enough to actually test it soon :)

+++ modules/core/src/main/java/at/redi2go/photonics/core/rendering/world/block/palette/PaletteTextureView.java
@@ -2,11 +2,12 @@ package at.redi2go.photonics.core.rendering.world.block.palette;
 
 import at.redi2go.photonics.api.Disposable;
 import org.joml.Vector4f;
+import org.joml.Vector4i;
 
 public interface PaletteTextureView extends Disposable {
     int pos();
 
-    void writeFace(int face, Vector4f value);
+    void writeFace(int face, Vector4i value);
 
     void upload();
 }
